+++ src/templates/blog-post.jsx
@@ -4,9 +4,11 @@
 import React from 'react';
 import { graphql } from 'gatsby';
 import styled from 'styled-components';
+import { ArrowLeftIcon, ArrowRightIcon } from '@primer/octicons-react';
 
 import Intl from '../components/Intl';
 import LayoutPost from '../components/layout/LayoutPost';
+import LocalizedLink from '../components/link/LocalizedLink';
 import Rule from '../components/Rule';
 import Theme from '../components/Theme';
 
@@ -34,8 +36,54 @@ const PostDiv = styled.div`
   }
 `;
 
+const NavLinks = styled.div`
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  column-gap: 2em;
+  margin: 2em ${(props) => props.theme.layout.marginPhone};
+
+  @media screen and ${(props) => props.theme.devices.phoneL} {
+    margin: 3em ${(props) => props.theme.layout.marginTablet};
+  }
+
+  @media screen and ${(props) => props.theme.devices.tablet} {
+    margin: 3em ${(props) => props.theme.layout.marginTablet};
+  }
+`;
+
+const Previous = styled.div`
+  word-break: break-all;
+  word-break: break-word;
+  hyphens: auto;
+`;
+
+const Next = styled.div`
+  word-break: break-all;
+  word-break: break-word;
+  hyphens: auto;
+  text-align: end;
+`;
+
+const ArrowLeft = styled(ArrowLeftIcon)`
+  margin: 0 0.2rem 0.1rem 0;
+`;
+
+const ArrowRight = styled(ArrowRightIcon)`
+  margin: 0 0 0.1rem 0.2rem;
+`;
+
 const Page = (props) => {
-  const { contentfulBlogPost: post } = props.data;
+  const { contentfulBlogPost: post, allContentfulBlogPost: posts } = props.data;
+
+  const previousEdges = posts.edges.filter(
+    ({ node }) => node.contentful_id === props.pageContext.pageId,
+  );
+  const nextEdges = posts.edges.filter(
+    ({ node }) => node.contentful_id === props.pageContext.pageId,
+  );
+
+  const previous = previousEdges.length > 0 ? previousEdges[0].previous : null;
+  const next = nextEdges.length > 0 ? nextEdges[0].next : null;
 
   return (
     <LayoutPost
@@ -43,11 +91,42 @@ const Page = (props) => {
       post={post}
       locale={props.pageContext.locale}
       pageId={props.pageContext.pageId}
+      description={post.body.childMarkdownRemark.excerpt}
     >
       <Separator>
         <Rule color="blue" mode={1} />
       </Separator>
       <PostDiv dangerouslySetInnerHTML={{ __html: post.body.childMarkdownRemark.html }} />
+      <NavLinks>
+        <Previous>
+          {(() => {
+            if (previous) {
+              return (
+                <>
+                  <ArrowLeft />
+                  <LocalizedLink to={previous.contentful_id} locale={props.pageContext.locale}>
+                    {previous.title}
+                  </LocalizedLink>
+                </>
+              );
+            }
+          })()}
+        </Previous>
+        <Next>
+          {(() => {
+            if (next) {
+              return (
+                <>
+                  <LocalizedLink to={next.contentful_id} locale={props.pageContext.locale}>
+                    {next.title}
+                  </LocalizedLink>
+                  <ArrowRight />
+                </>
+              );
+            }
+          })()}
+        </Next>
+      </NavLinks>
     </LayoutPost>
   );
 };
@@ -65,7 +144,12 @@ const BlogPost = (props) => (
 export default BlogPost;
 
 export const pageQuery = graphql`
-  query BlogPostQuery($pageId: String, $locale: String, $momentJsLocale: String) {
+  query BlogPostQuery(
+    $pageId: String
+    $locale: String
+    $momentJsLocale: String
+    $management: Boolean
+  ) {
     site {
       siteMetadata {
         simpleLocales {
@@ -85,12 +169,32 @@ export const pageQuery = graphql`
       }
       body {
         childMarkdownRemark {
+          excerpt
           html
         }
       }
       category {
         contentful_id
         name
+      }
+    }
+    allContentfulBlogPost(
+      filter: { management: { eq: $management }, node_locale: { eq: $locale } }
+      sort: { fields: date, order: DESC }
+    ) {
+      edges {
+        next {
+          contentful_id
+          title
+        }
+        node {
+          contentful_id
+          title
+        }
+        previous {
+          contentful_id
+          title
+        }
       }
     }
   }