+++ src/templates/blog.jsx
@@ -105,6 +105,7 @@ const Page = (props) => {
       title={i('blogTitle')}
       locale={props.pageContext.locale}
       pageId={props.pageContext.pageId}
+      description={props.data.contentfulIndexPage.description.description}
     >
       <Separator>
         <Rule color="peach" mode={2} />
@@ -195,6 +196,11 @@ export const pageQuery = graphql`
             name
           }
         }
+      }
+    }
+    contentfulIndexPage(node_locale: { eq: $locale }) {
+      description {
+        description
       }
     }
   }