+++ src/templates/management.jsx
@@ -123,7 +123,12 @@ const Page = (props) => {
   const { contentfulPage: page, allContentfulBlogPost: posts } = props.data;
 
   return (
-    <Layout title={page.title} locale={props.pageContext.locale} pageId={props.pageContext.pageId}>
+    <Layout
+      title={page.title}
+      locale={props.pageContext.locale}
+      pageId={props.pageContext.pageId}
+      description={page.description.description}
+    >
       <Div dangerouslySetInnerHTML={{ __html: page.body.childMarkdownRemark.html }} />
       <H2>{i('managementNewsTitle')}</H2>
       <Separator>
@@ -195,6 +200,9 @@ export const pageQuery = graphql`
         childMarkdownRemark {
           html
         }
+      }
+      description {
+        description
       }
     }
     allContentfulBlogPost(